+++ Worker/UnityMmo/Packages/games.cookiedragon.mmogf.core/MmogfMessages/Core/Components/Position.cs
@@ -6,8 +6,8 @@ namespace Mmogf.Core
     [MessagePackObject]
     public partial struct Position : IEntityComponent
     {
-        public static int ComponentId = 2;
-        public int GetComponentId() => ComponentId;
+        public static short ComponentId = 2;
+        public short GetComponentId() => ComponentId;
 
         [Key(0)]
         public double X { get; set; }
