+++ Worker/UnityMmo/Packages/games.cookiedragon.mmogf.core/MmogfMessages/Core/EntityInfo.cs
@@ -9,7 +9,7 @@ namespace Mmogf.Core
         [Key(0)]
         public int EntityId { get; set; }
         [Key(1)]
-        public Dictionary<int, byte[]> EntityData { get; set; }
+        public Dictionary<short, byte[]> EntityData { get; set; }
 
     }
 }
